moved plugin related commands to their own plugin
This is part 2 of #365 after #400.
